refactor: extract Ansi / Formatting / Styling into compilertop package
Splits CompilerTop's leaf helpers into a new
`ca.uwaterloo.flix.tools.compilertop` package as a first step toward
breaking up CompilerTop.scala. Pure code move, no behavior change:
ANSI escape codes + wrappers, text/numeric formatting helpers, and
threshold-driven `style*` functions each get their own file.
CompilerTop.scala shrinks by ~195 lines and now imports the new
modules with wildcard imports so call sites are unchanged.
